Develop: dynamic data for gaugeChart
