+++ src/view/PlayerDetail/PlayerDetailPage.vue
@@ -1,25 +1,128 @@
 <script setup>
+import { nextTick } from 'vue';
 import { storeToRefs } from "pinia";
 import { useRoute } from 'vue-router';
-import GaugeChart from './components/gaugeChart.vue';
 import RadarChart from './components/RadarChart.vue';
 import { useAppStore } from '../../stores/appStore';
 import { usePlayerStore } from "../../stores/playerStore.js";
+import * as echarts from 'echarts';
 
-const { player_data } = usePlayerStore();
+const gauge = ref();
+let gaugeChart;
+const { player_data, normalizedRating } = usePlayerStore();
 const appStore = useAppStore();
 const { sideStatus } = storeToRefs(appStore);
 const route = useRoute();
 const value1 = ref(3)
 const { id } = route.params;
 const { stats } = player_data[id];
+const contributeRating = ref(0);
+
 
 
 onMounted(() => {
     sideStatus.value = true;
+    const playerKey = `${player_data[id].position.toUpperCase()}S`;
+    normalizedRating[playerKey].forEach(player => {
+        if (player.name === player_data[id].name) {
+            contributeRating.value = player.normalizedRating * 0.01;
+        }
+    });
+
+    console.log(contributeRating.value);
 });
 
+nextTick(() => {
+    const option = {
+        series: [
+            {
+                type: 'gauge',
+                startAngle: 180,
+                endAngle: 0,
+                center: ['50%', '75%'],
+                radius: '90%',
+                min: 0,
+                max: 1,
+                splitNumber: 8,
+                axisLine: {
+                    lineStyle: {
+                        width: 6,
+                        color: [
+                            [0.25, '#FF6E76'],
+                            [0.5, '#FDDD60'],
+                            [0.75, '#58D9F9'],
+                            [1, '#7CFFB2']
+                        ]
+                    }
+                },
+                pointer: {
+                    icon: 'path://M12.8,0.7l12,40.1H0.7L12.8,0.7z',
+                    length: '12%',
+                    width: 20,
+                    offsetCenter: [0, '-60%'],
+                    itemStyle: {
+                        color: 'auto'
+                    }
+                },
+                axisTick: {
+                    length: 12,
+                    lineStyle: {
+                        color: 'auto',
+                        width: 2
+                    }
+                },
+                splitLine: {
+                    length: 20,
+                    lineStyle: {
+                        color: 'auto',
+                        width: 5
+                    }
+                },
+                axisLabel: {
+                    color: '#464646',
+                    fontSize: 20,
+                    distance: -60,
+                    rotate: 'tangential',
+                    formatter: function (value) {
+                        if (value === 0.875) {
+                            return 'A';
+                        } else if (value === 0.625) {
+                            return 'B';
+                        } else if (value === 0.375) {
+                            return 'C';
+                        } else if (value === 0.125) {
+                            return 'D';
+                        }
+                        return '';
+                    }
+                },
+                title: {
+                    offsetCenter: [0, '-10%'],
+                    fontSize: 20
+                },
+                detail: {
+                    fontSize: 30,
+                    offsetCenter: [0, '-35%'],
+                    valueAnimation: true,
+                    formatter: function (value) {
+                        return Math.round(value * 100) + '';
+                    },
+                    color: 'inherit'
+                },
+                data: [
+                    {
+                        value: contributeRating.value,
+                        name: 'Contribution'
+                    }
+                ]
+            }
+        ]
+    };
+    gaugeChart = echarts.init(gauge.value);
+    gaugeChart.setOption(option);
+});
 </script>
+
 <template>
     <div class="player-detail_header">
         <div class="player-detail-content">
@@ -46,12 +149,12 @@ onMounted(() => {
             <img :src='player_data[id].detail_image' alt="player" />
         </div>
         <div class="chart-group">
-            <GaugeChart />
+            <div ref="gauge" class="gauge-chart"></div>
+
             <RadarChart />
         </div>
     </div>
 </template>
-
 
 <style lang="scss" scoped>
 .player-detail_header {
@@ -105,6 +208,11 @@ onMounted(() => {
         display: flex;
         justify-content: center;
         flex-direction: column;
+
+        .gauge-chart {
+            width: 35vw;
+            height: 40vh;
+        }
     }
 }
